+++ constants/Colors.js
@@ -18,4 +18,5 @@ export default {
   indicatorYellow: "#FFB800",
   indicatorLime: "#A6BC35",
   indicatorGreen: "#6DD149",
+  linkBlue: "#2980b9",
 };
