+++ lib/WebBrowserManager.js
@@ -1,0 +1,7 @@
+import { WebBrowser } from "expo";
+
+const openLink = url => WebBrowser.openBrowserAsync(url);
+
+export default {
+  openLink,
+};
